+++ CHANGELOG.md
@@ -40,6 +40,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 * Add CLI commands for the exchange module endpoints and queries [#1701](https://github.com/provenance-io/provenance/issues/1701).
+* Create CLI commands for adding a market to a genesis file [#1757](https://github.com/provenance-io/provenance/issues/1757).
+
+### Improvements
+
+* Create a default market in `make run`, `localnet`, `devnet` and the `provenanced testnet` command [#1757](https://github.com/provenance-io/provenance/issues/1757).
 
 ---